requirements.txt
asgiref==3.9.1
black==25.9.0
click==8.3.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
gunicorn==23.0.0
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
pytokens==0.2.0
sqlparse==0.5.3
